+++ R/add-pathways.R
@@ -1,24 +1,26 @@
 #' Pathway Alterations
 #'
-#' Input a binary matrix of patients x genes and return a dataframe with a column per pathway
+#' Input a binary matrix of patients x alterations and return a dataframe with a column per pathway
 #' indicating if default or custom oncogenic signaling pathways
-#' are activated in each sample. Pathways were curated
+#' are activated in each sample. Default package pathways were sourced
 #' from [Sanchez-Vega, F et al., 2018](https://pubmed.ncbi.nlm.nih.gov/29625050/).
 #'
-#' Please check for gene aliases in your dataset before using.
+#' Please check for gene aliases in your data set before using.
 #'
 #'
-#' @param gene_binary a binary matrix from `gene_binary()`
-#' @param pathways a vector of pathway names to annotate. The options are `names(gnomeR::pathways)` ("RTK/RAS", "Nrf2",
-#'  "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch"). You can pass one pathway name, multiple pathway names, or `NULL`. By default, all
-#'  pathways defined in `gnomeR::pathways` will be included. Included default pathways are alteration-specific meaning a specific type of alteration (mut/cna/fusion)
-#'  is required to mark a 1 for that pathway. If you want gene-level pathways, use `summarize_by_gene()` on your binary matrix, then add
-#'  pathways you want as `custom_pathways`
-#' @param custom_pathways a vector of alterations to annotate as a pathway, or a list of custom pathways (see `gnomeR::pathways` as example)
-#' @param count_pathways_by Must be one of the following: "alteration" (default), or "gene". This determines whether
-#' any type of gene alteration should be counted towards a pathway ("gene") or only specific types of alterations should be counted towards a pathway ("alteration")
-#' By default, the function assumes alteration-specific pathway annotation and all default pathways are annotated this way. If a
-#' custom pathway is passed with no suffix (e.g. `custom_pathway = 'TP53'`) it will assume it is a mutation.
+#' @description
+#' Please note that only `sample_id column`, and columns with .Amp, .Del, .fus or no suffix are accepted.
+#' Any gene column with no suffix will be assumed to be a mutation.
+#'
+#' @param gene_binary a binary matrix from `create_gene_binary()`
+#' @param pathways a vector of pre-coded pathways to annotate. The options are `names(gnomeR::pathways)` ("RTK/RAS", "Nrf2",
+#'  "PI3K", "TGFB", "p53", "Wnt", "Myc", "Cell cycle", "Hippo", "Notch"). You can pass multiple pathway names, or `NULL`. By default, all
+#'  pathways defined in `gnomeR::pathways` will be included. Included default pathways are alteration-specific, meaning a specific type of alteration (mut/cna/fusion)
+#'  is required to mark a 1 for that pathway.
+#' @param custom_pathways a vector of alterations to annotate as a single pathway, or a list of custom pathways (see `gnomeR::pathways` as example).
+#' You must specify the alteration type for each gene using `.mut`, `.Amp`, `.Del` suffix, e.g. `c("TP53.mut", "CDKN2A.Amp")`. If you wish to count any type of
+#' alteration on that gene towards the pathway you can use the `.any` suffix (e.g. `c("TP53.any")`).
+#' @param count_pathways_by deprecated
 #' @keywords internal
 #' @return a data frame: each sample is a row, columns are pathways, with values of 0/1 depending on pathway alteration status.
 #' @export
@@ -34,20 +36,28 @@
 add_pathways <- function(gene_binary,
                          pathways = c(names(gnomeR::pathways)),
                          custom_pathways = NULL,
-                         count_pathways_by = c("alteration", "gene")) {
+                         count_pathways_by = deprecated()) {
+
+  # Check arguments -----------------------------------------------------------
 
   all_path <- gnomeR::pathways
   all_path_names <- names(all_path)
-
-  # check arguments -----------------------------------------------------------
-
-  # custom pathways
-  switch(!(class(custom_pathways) %in% c("NULL", "character", "list")),
-         cli::cli_abort("{.code custom_pathways} must be character vector, or list"))
-
   .check_required_cols(gene_binary, "sample_id", "gene_binary")
 
-  # user-specified pathways
+  # * Deprecated Arguments (will remove this in the future) ----
+
+  if (lifecycle::is_present(count_pathways_by)) {
+    lifecycle::deprecate_stop(when = "1.3.0",
+                              what = "add_pathways(count_pathways_by)",
+                              details = c(
+                                i = c("All pathways are now counted on the specific alteration level, not the gene level and all columns in your data with no suffix (e.g. `data$TP53`) are assumed to be mutations.
+                                You must explicitely specify in your custom pathway what types of alterations count towards pathway e.g. `custom_pathway = c('TP53.mut', 'APC.Del').
+                                      If you would like to count any type of alteration on a gene towards a pathway, use the suffix `.all`, e.g. `custom_pathway = c('TP53.all', 'APC.all')")))
+  }
+
+
+
+  # * Default pathways ----
   pathways_input <- pathways
 
   if(!is.null(pathways)) {
@@ -59,16 +69,12 @@ add_pathways <- function(gene_binary,
   switch(length(not_valid) > 0,
          cli::cli_warn("Ignoring {.code {not_valid}}: not a known pathway. See {.code gnomeR::pathways}"))
 
-  # count pathways by
-  count_pathways_by <- match.arg(count_pathways_by, c("alteration", "gene"))
 
-  all_cols <- colnames(gene_binary)
-  mut_cols <- !(str_detect(all_cols, ".Amp|.Del|.fus|.cna"))
+  # Custom pathways ----
+  switch(!(class(custom_pathways) %in% c("NULL", "character", "list")),
+         cli::cli_abort("{.code custom_pathways} must be character vector, or list"))
 
-  switch(is.null(custom_pathways) & count_pathways_by == "gene",
-         cli::cli_alert_warning("Annotating the default pathways by gene may be inappropriate."))
 
-  # custom_pathways:  can be list or vector------------------------------------
   if (!is.null(custom_pathways)) {
 
     # if vector, turn into a list---
@@ -90,57 +96,52 @@ add_pathways <- function(gene_binary,
       }
     }
 
+    if(any(purrr::map_lgl(custom_pathways, ~any(!str_detect(.x, ".Amp|.Del|.fus|.mut"))))) {
+      cli::cli_abort(c("All alterations specified in {.code custom_pathway} must have one ",
+                      "of the following suffixes specified: {.code .Amp}, {.code .Del}, {.code .fus}, {.code .mut}, {.code .all}"))
+    }
 
-    if(count_pathways_by == "alteration") {
 
-      # add mut on custom pathways when count_pathways_by == "alteration"
-      if(any(purrr::map_lgl(custom_pathways, ~any(!str_detect(.x, ".Amp|.Del|.fus|.cna|.mut"))))) {
-        cli::cli_inform("Assuming any gene in {.code custom_pathway} without
-        suffix {.code .Amp|.Del|.fus|.cna} is specifically a mutation in that pathway. CNA and fusions will not be counted (e.g. TP53.Del). To control this behavior, see argument {.code count_pathways_by}")
+    # process GENE.all suffix in custom pathways
+    custom_pathways <- purrr::map(custom_pathways, function(x) {
+      gene_all <- x[str_detect(x, ".all")]
+
+      if(length(gene_all) > 0) {
+        gene_all <-  str_remove(gene_all, ".all")
+        gene_add <- paste0(gene_all, c(".Amp", ".Del", ".fus", ".mut"))
+        x <- c(x[!str_detect(x, ".all")], gene_add)
       }
 
+      unique(x)
 
-      custom_pathways <- purrr::map(custom_pathways, function(x) {
-        x[!str_detect(x, ".Amp|.Del|.fus|.cna")] <-
-          paste0(stringr::str_trim(
-            x[!str_detect(x, ".Amp|.Del|.fus|.cna")]), ".mut")
-
-        unique(x)
-
-        } )
+      } )
 
     }
-
-    if(count_pathways_by == "gene" &
-       any(purrr::map_lgl(custom_pathways, ~any(str_detect(.x, ".Amp|.Del|.fus|.cna"))))) {
-      cli::cli_warn("You selected {.code count_pathways_by = 'gene'}. Ignoring {.code .Amp|.Del|.fus|.cna} in {.code custom_pathway} genes passed")
-    }
-
-    }
-
 
 
   # get user selected pathways (if any)
   pathways <- all_path[pathways]
   final_paths <- c(custom_pathways, pathways)
 
-  if(count_pathways_by == "gene") {
-    final_paths <- purrr::map(final_paths, ~unique(str_remove_all(.x, ".mut|.Amp|.Del|.fus|.cna|.meth")))
-  }
-
   # prep data ------------------------------------------------------------------
+  all_cols <- colnames(gene_binary)
+  mut_cols <- !(str_detect(all_cols, ".Amp|.Del|.fus"))
 
   # rename mut cols -assume all non CNA/Fusion are mutations
   if (any(mut_cols)) {
+
+    # in case any columns already had .mut suffix
+    all_cols[mut_cols] <- str_remove(all_cols[mut_cols], ".mut")
+
     all_cols[mut_cols] <- paste0(all_cols[mut_cols], ".mut")
+
     colnames(gene_binary) <- all_cols
   }
 
   # process pathways ---------------------------------------------------------------
   path_out <- purrr::imap_dfc(final_paths, ~ .sum_alts_in_pathway(gene_binary = gene_binary,
                                                                  pathway_list_item = .x,
-                                                                 pathway_name = .y,
-                                                                 count_pathways_by = count_pathways_by))
+                                                                 pathway_name = .y))
 
   # return data  ---------------------------------------------------------------
 
@@ -151,7 +152,6 @@ add_pathways <- function(gene_binary,
   }
 
   path_out <- gene_binary %>%
-#    select("sample_id") %>%
     bind_cols(path_out)
 
   return(path_out)
@@ -164,7 +164,6 @@ add_pathways <- function(gene_binary,
 #' @param pathway_list_item a named list of length 1 with pathway name as name and vector of genes as first
 #' and only item in list
 #' @param pathway_name name of pathway
-#' @param count_pathways_by passed from `add_pathways()`
 #'
 #' @return a dataframe of 1 column of 0/1s indicating pathway activated yes/no
 #' @keywords internal
@@ -181,15 +180,8 @@ add_pathways <- function(gene_binary,
 .sum_alts_in_pathway <- function(gene_binary, pathway_list_item,
                                  pathway_name,
                                  count_pathways_by) {
-  path_alt <- switch(count_pathways_by,
-                     "alteration" = {
-                       gene_binary %>%
-                         select(any_of(unlist(pathway_list_item, use.names=FALSE)))
-                     },
-                     "gene" = {
-                       gene_binary %>%
-                         select(contains(unlist(pathway_list_item, use.names=FALSE)))
-                     }) %>%
+  path_alt <- gene_binary %>%
+    select(any_of(unlist(pathway_list_item, use.names=FALSE))) %>%
     mutate(sum = rowSums(., na.rm = TRUE)) %>%
     transmute('pathway_{pathway_name}' := if_else(sum >= 1, 1, 0))